Creating base training router
